mp:
 - implement the mp code biases estimator in the form
   of a trait
 - add documentation
