+++ rinex-cli/doc/processing.md
@@ -21,6 +21,11 @@ This tool supports standard GNSS recombinations, especially for modern RINEX.
 Refer to this page for [thorough documentation](gnss-combination.md).  
 It is important to understand GNSS signal recombinations and what they can represent.  
 
+The RINEX tool suites can estimate the following code biases:
+
+- the DCB code biases 
+- the MP code biases
+
 Differential Code Biases (DBCs)
 ===============================
 
@@ -55,20 +60,9 @@ rinex-cli \
 
 <img align="center" width="650" src="https://github.com/gwbres/rinex/blob/main/doc/plots/esbc00dnk_ph_dcbs.png">
 
-Differential Processing
-=======================
 
-When moving to more advanced RINEX processing,
-Navigation RINEX (Ephemeris) must be provided with `--nav`.
-
-In this mode, `--fp` is expected to be an Observation RINEX.
-
-Let's remind the user that in this mode, `--sv-epoch` helps
-exhibit which vehicles share Ephemeris and Observations for a given epoch
-or epoch range. This feature is very important do determine
-which vehicle is a good candidate for the operations that follow.
-
-## Code Multipath (MP) analysis
+Code Multipath biases
+=====================
 
 MP ratios (so called "CMC" for Code Minus Carrier metrics) 
 are formed by combining Phase and PR observations sampled against different carrier frequencies
@@ -83,17 +77,32 @@ A very compelling use case of MP code analysis
 can be
 [found here](https://www.taoglas.com/wp-content/uploads/pdf/Multipath-Analysis-Using-Code-Minus-Carrier-Technique-in-GNSS-Antennas-_WhitePaper_VP__Final-1.pdf).
 
+
+MP analysis is summoned with `--mp`.
+
 In this example, `ESBC00DNK_R_2020` vehicle GPS#13 
-has enough data to compute MP ratio for codes "1C", "2W" and "5Q".
-This operation requires combining the associated Ephemeris, that we also provide
-for demonstration purposes:
+has enough data to compute MP ratios:
 
 ```bash
 rinex-cli \
     --fp test_resources/CRNX/V3/ESBC00DNK_R_20201770000_01D_30S_MO.crx.gz \
-    --nav test_resources/NAV/V3/ESBC00DNK_R_20201770000_01D_MN.rnx.gz \
-    -P G13 --mp
+        -P G13 --dcb --mp
 ```
+
+<img align="center" width="650" src="https://github.com/gwbres/rinex/blob/main/doc/plots/esbc00dnk_g13_dcb_mp.png">
+
+Differential Processing
+=======================
+
+When moving to more advanced RINEX processing,
+Navigation RINEX (Ephemeris) must be provided with `--nav`.
+
+In this mode, `--fp` is expected to be an Observation RINEX.
+
+Let's remind the user that in this mode, `--sv-epoch` helps
+exhibit which vehicles share Ephemeris and Observations for a given epoch
+or epoch range. This feature is very important do determine
+which vehicle is a good candidate for the operations that follow.
 
 ## Cycle slips analysis
